Adjusted models and their direct tests to be namespace-compatible.
