+++ common/predictive-text/unit_tests/headless/worker-predict-dummy.js
@@ -4,6 +4,7 @@
 
 var assert = require('chai').assert;
 
+debugger
 var DummyModel = require('../../build/intermediate').models.DummyModel;
 
 describe('LMLayerWorker dummy model', function() {
